requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==2.1.1
discord.py==2.1.0
frozenlist==1.3.3
idna==3.4
lxml==4.9.2
multidict==6.0.4
python-dotenv==0.21.0
requests==2.28.2
schedule==1.1.0
soupsieve==2.3.2.post1
urllib3==1.26.14
yarl==1.8.2
